simplify arraylib, add 3.12 to ci
